the alert is working

+++ js/controllers/recipecontroller.js
@@ -1,9 +1,11 @@
 App.RecipeController = Ember.ObjectController.extend({
   actions: {
     destroy: function () {
-     this.get('model').deleteRecord();
+     if (confirm('Once you delete, it\'s gone forever!')){
+       this.get('model').deleteRecord();
      this.get('model').save();
      this.transitionToRoute('index');
-        },
-    },
+        }
+      }
+    }
 });
